[OPIK-1373] guardrail chart
