+++ apps/opik-frontend/src/api/projects/useProjectMetric.ts
@@ -8,6 +8,7 @@ export enum METRIC_NAME_TYPE {
   DURATION = "DURATION",
   TOKEN_USAGE = "TOKEN_USAGE",
   COST = "COST",
+  FAILED_GUARDRAILS = "FAILED_GUARDRAILS",
 }
 
 export enum INTERVAL_TYPE {
